tratare exceptii + autorizare endpointuri

+++ Tema_dotNet.Database/Repositories/ProducatorRepository.cs
@@ -32,18 +32,38 @@ namespace Tema_dotNet.Database.Repositories
 
         public void Add(Producator producator)
         {
+            if (_context.Producatori.Any(p => p.Nume == producator.Nume))
+            {
+                throw new Exception($"Producator with name {producator.Nume} already exists");
+            }
             _context.Producatori.Add(producator);
             _context.SaveChanges();
         }
 
         public void Update(Producator producator)
         {
+            if (producator == null)
+            {
+                throw new Exception("Producator not found");
+            }
+            if (_context.Producatori.Any(p => p.Nume == producator.Nume && p.Id != producator.Id))
+            {
+                throw new Exception($"Producator with name {producator.Nume} already exists");
+            }
             _context.Producatori.Update(producator);
             _context.SaveChanges();
         }
 
         public void Delete(Producator producator)
         {
+            if (producator == null)
+            {
+                throw new Exception("Producator not found");
+            }
+            if (producator.Produse.Any())
+            {
+                throw new Exception("Producator has products");
+            }
             _context.Producatori.Remove(producator);
             _context.SaveChanges();
         }
