+++ Tema_dotNet.Database/Repositories/ProdusRepository.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc.Filters;
 using Tema_dotNet.Database.Context;
 using Tema_dotNet.Database.Entities;
 
@@ -31,12 +32,24 @@ namespace Tema_dotNet.Database.Repositories
 
         public void AddProdus(Produs produs)
         {
+            if (_context.Produse.Any(p => p.Nume == produs.Nume))
+            {
+                throw new Exception($"Produs with name {produs.Nume} already exists");
+            }
             _context.Produse.Add(produs);
             _context.SaveChanges();
         }
 
         public void EditProdus(Produs produs, Produs payload)
         {
+            if (produs == null)
+            {
+                throw new Exception("Produs not found");
+            }
+            if (_context.Produse.Any(p => p.Nume == payload.Nume && p.Id != payload.Id))
+            {
+                throw new Exception($"Produs with name {payload.Nume} already exists");
+            }
             produs.Nume = payload.Nume;
             produs.Pret = payload.Pret;
             produs.ProducatorId = payload.ProducatorId;
@@ -45,6 +58,10 @@ namespace Tema_dotNet.Database.Repositories
 
         public void DeleteProdus(Produs produs)
         {
+            if (produs == null)
+            {
+                throw new Exception("Produs not found");
+            }
             _context.Produse.Remove(produs);
             _context.SaveChanges();
         }
